requirements.txt
python-telegram-bot==20.7
matplotlib==3.8.2
pandas==2.1.4
python-dotenv==1.0.0
pytz==2024.1
schedule==1.2.1
